Refactor host validation to avoid config imports
